modify BaseStateHandler class
I put in init two parameters: viewer, and controller. Both should put
argument classes which a implementation of BaseViewer and BaseController
respectively.

Also I add back view method to handle rendering which will be delegated
to the BaseVieawer implementer. And control method to handle update and
controls which will be delgated to controller class.