fix highlighting bug using update viewDelegate method
In this case I am using the viewDelegate to pass the value and then also
modifying local value using getter and setter.

This is called delegate model which using the instance of self (the
instance of the StartView) to act as self.viewDelegate which is passed
as the parameter argument each time update function in
StartViewController is called.