Use controls tables to set the control keys
I decided to use table called self.contols to list all controls on the
paddle class. This should make it easier in the future if I want to
alter any keys related to any control movements.

ADDITION: I also put escape key feature in the paused state to enable
the developer during this stage to be able to exit from the app even in
paused state.