ServeViewController diagram
I already put the class in the diagram but then I realize something.
Each View (palyView and StartView) all init refer to BaseController
which is the parent class of each those view controllers.

This is wrong! I need to fix this!

TODO: put issue about this make each view directly pointed to each
controller and only the controller should point to the superclass